Added passing. Fixed Draw Card glitch. Fixed wild color change glitch

+++ Uno/GameObjects/Deck.h
@@ -43,7 +43,7 @@ private:
 
 struct RNG {
     int operator() (int n) {
-        return std::rand() / (1.0 + RAND_MAX) * n;
+        return std::rand() / (int)(1.0 + RAND_MAX) * n;
     }
 };
